Moved tests to appropriate folders
