+++ test/service/ProjectServiceTest.java
@@ -1,16 +1,20 @@
-package application;
+package service;
 
 import integration.DatabaseTest;
 import models.*;
 import org.junit.Test;
 import services.DonationGoalService;
+import services.ProjectService;
+
 import java.sql.Date;
+import java.util.ArrayList;
+import java.util.List;
 
 import static org.junit.Assert.assertEquals;
 
-public class DonationGoalServiceTest {
-
-    public Project getFilledDataStructure() {
+public class ProjectServiceTest {
+    public List<Project> getFilledDataStructure() {
+        List<Project> projectList = new ArrayList<>();
         User user = new User();
         user.save();
 
@@ -22,19 +26,19 @@ public class DonationGoalServiceTest {
         type2.setName("Hemden");
         type2.save();
 
-        Project project = new Project();
-        project.save();
+        Project project1 = new Project();
+        project1.save();
 
         DonationGoal goal1 = new DonationGoal();
         goal1.setAmount(10);
         goal1.setType(type1);
-        goal1.setProject(project);
+        goal1.setProject(project1);
         goal1.save();
 
         DonationGoal goal2 = new DonationGoal();
-        goal2.setAmount(20);
+        goal2.setAmount(30);
         goal2.setType(type2);
-        goal2.setProject(project);
+        goal2.setProject(project1);
         goal2.save();
 
         Donation donation1 = new Donation();
@@ -54,32 +58,53 @@ public class DonationGoalServiceTest {
         donation2.save();
 
         Donation donation3 = new Donation();
-        donation3.setAmount(20);
+        donation3.setAmount(5);
         donation3.setDate(new Date(5464));
         donation3.setDonationGoal(goal2);
         donation3.setUser(user);
         donation3.setMessageToCollector("123456");
         donation3.save();
 
-        project.refresh();
+        Project project2 = new Project();
+        project2.save();
 
-        return project;
+        DonationGoal goal3 = new DonationGoal();
+        goal3.setAmount(10);
+        goal3.setType(type1);
+        goal3.setProject(project2);
+        goal3.save();
+
+        Donation donation4 = new Donation();
+        donation4.setAmount(12);
+        donation4.setDate(new Date(5464));
+        donation4.setDonationGoal(goal3);
+        donation4.setUser(user);
+        donation4.setMessageToCollector("123456");
+        donation4.save();
+
+        project1.refresh();
+        project2.refresh();
+        projectList.add(project1);
+        projectList.add(project2);
+
+        return projectList;
     }
+
     @Test
-    public void testGetStateWithSingleDonation() {
+    public void testStateSmallerThanGoal() {
         DatabaseTest.runInCleanApp(testBrowser -> {
-            Project project = getFilledDataStructure();
-            int state1 = DonationGoalService.getState(project.getDonationGoals().get(0));
+            List<Project> projectList = getFilledDataStructure();
+            int state1 = ProjectService.getStateOfProjectInPercent(projectList.get(0));
             assertEquals(50, state1);
         });
     }
 
     @Test
-    public void testGetStateWithMultipleDonations() {
+    public void testStateBiggerThanGoal() {
         DatabaseTest.runInCleanApp(testBrowser -> {
-            Project project = getFilledDataStructure();
-            int state2 = DonationGoalService.getState(project.getDonationGoals().get(1));
-            assertEquals(100, state2);
+            List<Project> projectList = getFilledDataStructure();
+            int state1 = ProjectService.getStateOfProjectInPercent(projectList.get(1));
+            assertEquals(100, state1);
         });
     }
 }
